requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aniso8601==9.0.1
    # via graphene
asgiref==3.8.1
    # via
    #   django
    #   django-axes
    #   django-cors-headers
authlib==1.3.2
    # via drf-oidc-auth
cachetools==5.5.0
    # via django-helusers
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==43.0.1
    # via
    #   authlib
    #   drf-oidc-auth
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
deprecation==2.1.0
    # via django-helusers
django==4.2.16
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   django-graphql-jwt
    #   django-helusers
    #   django-ilmoitin
    #   django-mailer
    #   django-parler
    #   djangorestframework
    #   drf-oidc-auth
    #   graphene-django
    #   helsinki-profile-gdpr-api
    #   social-auth-app-django
django-admin-list-filter-dropdown==1.0.3
    # via -r requirements.in
django-anymail==12.0
    # via django-ilmoitin
django-axes==7.0.0
    # via -r requirements.in
django-cors-headers==4.4.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-graphql-jwt==0.4.0
    # via -r requirements.in
django-helusers==0.13.0
    # via
    #   -r requirements.in
    #   helsinki-profile-gdpr-api
django-ilmoitin==0.7.0
    # via -r requirements.in
django-mailer==2.3.2
    # via django-ilmoitin
django-parler==2.3
    # via django-ilmoitin
django-utils-six==2.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   drf-oidc-auth
    #   helsinki-profile-gdpr-api
drf-oidc-auth==3.0.0
    # via
    #   -r requirements.in
    #   helsinki-profile-gdpr-api
ecdsa==0.19.0
    # via python-jose
factory-boy==3.3.1
    # via -r requirements.in
faker==30.3.0
    # via factory-boy
geographiclib==2.0
    # via geopy
geopy==2.4.1
    # via -r requirements.in
graphene==3.3
    # via
    #   django-graphql-jwt
    #   graphene-django
graphene-django==3.2.2
    # via
    #   -r requirements.in
    #   django-graphql-jwt
graphene-file-upload==1.3.0
    # via -r requirements.in
graphql-core==3.2.4
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
helsinki-profile-gdpr-api==0.2.0
    # via -r requirements.in
idna==3.10
    # via requests
importlib-metadata==8.5.0
    # via -r requirements.in
jinja2==3.1.4
    # via django-ilmoitin
lockfile==0.12.2
    # via django-mailer
markupsafe==3.0.1
    # via jinja2
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.1
    # via deprecation
promise==2.3
    # via graphene-django
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via
    #   django-graphql-jwt
    #   social-auth-core
python-dateutil==2.9.0.post0
    # via faker
python-jose==3.3.0
    # via django-helusers
python3-openid==3.2.0
    # via social-auth-core
pytz==2024.2
    # via -r requirements.in
requests==2.32.3
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-helusers
    #   drf-oidc-auth
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rsa==4.9
    # via python-jose
sentry-sdk==2.16.0
    # via -r requirements.in
six==1.16.0
    # via
    #   ecdsa
    #   graphene-file-upload
    #   promise
    #   python-dateutil
social-auth-app-django==5.4.2
    # via -r requirements.in
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.1
    # via django
text-unidecode==1.3
    # via graphene-django
typing-extensions==4.12.2
    # via faker
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   django-anymail
    #   requests
    #   sentry-sdk
zipp==3.20.2
    # via importlib-metadata
